Various fixes (#48)

+++ Scripts/functions.ps1
@@ -77,15 +77,18 @@ function Invoke (
 }
 
 function RemoveResourceGroups (
-    [parameter(Mandatory=$false)][object[]]$resourceGroups
+    [parameter(Mandatory=$false)][object[]]$resourceGroups,
+    [parameter(Mandatory=$false)][bool]$Force=$false
 ) {
     if ($resourceGroups) {
         $resourceGroupNames = $resourceGroups | Select-Object -ExpandProperty ResourceGroupName
-        Write-Host "If you wish to proceed removing these resource groups:`n$resourceGroupNames `nplease reply 'yes' - null or N aborts" -ForegroundColor Blue
-        $proceedanswer = Read-Host
-        if ($proceedanswer -ne "yes") {
-            Write-Host "`nSkipping $resourceGroupNames" -ForegroundColor Yellow
-            continue
+        if (!$Force) {
+            Write-Host "If you wish to proceed removing these resource groups:`n$resourceGroupNames `nplease reply 'yes' - null or N aborts" -ForegroundColor Blue
+            $proceedanswer = Read-Host
+            if ($proceedanswer -ne "yes") {
+                Write-Host "`nSkipping $resourceGroupNames" -ForegroundColor Yellow
+                return $false
+            }
         }
         $resourceGroups | Remove-AzResourceGroup -AsJob -Force
         return $true
